WIP | adding data_controller node
