+++ src/data_controller.cpp
@@ -474,21 +474,3 @@ namespace quirkd {
   }
 
 } // namespace quirkd
-
-int main(int argc, char** argv)
-{
-  ros::init(argc, argv, "DataController");
-  ros::NodeHandle nh;
-  quirkd::DataController dp(nh);
-  ros::Rate r(30);
-
-  dp.update();
-  while (ros::ok())
-  {
-    ros::spinOnce();
-    dp.update();
-    r.sleep();
-  }
-  ROS_INFO("DataController Exited.");
-  return 0;
-}
